+++ src/account/types/mod.rs
@@ -24,7 +24,10 @@ use serde::{Deserialize, Deserializer, Serialize};
 
 pub use self::{
     address::{AccountAddress, AddressWithUnspentOutputs},
-    balance::{AccountBalance, BaseCoinBalance, NativeTokensBalance},
+    balance::{
+        AccountBalance, AccountBalanceDto, BaseCoinBalance, BaseCoinBalanceDto, NativeTokensBalance,
+        NativeTokensBalanceDto, RequiredStorageDeposit,
+    },
 };
 use crate::account::Account;