+++ CHANGELOG.md
@@ -19,6 +19,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 1.0.0-rc.4 - 20XX-XX-XX
+
+### Added
+
+- `RequiredStorageDeposit` and `RequiredStorageDepositDto` types;
+- `account/types/balance` module;
+
+### Changed
+
+- `AccountBalance::required_storage_deposit` changed from `u64` to `RequiredStorageDeposit`;
+- `AccountBalanceDto::required_storage_deposit` changed from `String` to `RequiredStorageDepositDto`;
+- Move all balance related types to the `account/types/balance` module;
+- `AccountBalanceDto`, `BaseCoinBalanceDto` and `NativeTokensBalanceDto` moved from `message_interface/dtos` to `account/types/balance`;
+
 ## 1.0.0-rc.3 - 2022-11-24
 
 ### Added